stock market modified

+++ src/main/java/edu/bu/homeworkteam/bankatm/entities/Shareholding.java
@@ -8,5 +8,5 @@ import javax.persistence.Embeddable;
 @Data
 public class Shareholding {
     private int numberOfShares;
-    private int averageCostPricePerShare;
+    private float averageCostPricePerShare;
 }
